First attempt at a client that submits data to pastebin.
Currently this doesn't work alongside the JS version. The submission works (which is expected; the server doesn't understand the submission) but decrypting in browser fails.

The error message is "INVALID: invalid aes key size" which I think means that the key I'm using to encrypt in Python SJCL is not acceptable to Ecmascript SJCL. I see that ks=128 in the submission but I don't think JS SJCL uses that to obtain the key size :)